+++ src/ImplementationV2.sol
@@ -3,13 +3,13 @@ pragma solidity ^0.8.13;
 
 contract ImplementationV2 {
     address private owner;
-    uint public num;
+    uint256 public num;
 
-    function setNum(uint _num) public {
+    function setNum(uint256 _num) public {
         num = _num;
     }
 
-    function putNum(uint _num) public {
+    function putNum(uint256 _num) public {
         num = 2 * _num;
     }
 }
