+++ src/Proxy.sol
@@ -4,8 +4,9 @@ pragma solidity ^0.8.13;
 import "../lib/openzeppelin-contracts/contracts/access/Ownable.sol";
 
 contract Proxy is Ownable {
-    uint public num;
+    uint256 public num;
     address public implementation;
+
     event FallbackCall();
     event SetImplementationCall();
 
@@ -19,7 +20,7 @@ contract Proxy is Ownable {
     }
 
     fallback() external {
-        (bool success, ) = implementation.delegatecall(msg.data);
+        (bool success,) = implementation.delegatecall(msg.data);
 
         if (!success) {
             revert();
